requirements.txt
websockets>=12.0
pandas>=2.0.0
numpy>=1.24.0
binance-connector>=3.1.0
Flask>=3.0.0
psutil>=5.9.0